src/ai-service/requirements.txt
# CatVTON Dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
huggingface-hub>=0.20.0
